requirements.txt
yfinance>=0.2.36
supabase==2.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
